requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
moviepy==1.0.3
pydub==0.25.1
noisereduce==3.0.2
openai-whisper
transformers==4.44.2
torch==2.3.0
torchaudio==2.3.0
silero-vad==5.1.2
nltk==3.8.1
underthesea==6.8.4
psutil==6.0.0
schedule==1.2.2
scipy==1.14.1